Added permissions checking to ctnr_detail view

+++ cyder/core/ctnr/views.py
@@ -114,17 +114,27 @@ def create_user_extra_cols(ctnr, ctnrusers):
 
 
 def remove_user(request, ctnr_pk, user_pk):
-    CtnrUser.objects.get(ctnr_id=ctnr_pk, user_id=user_pk).delete()
-    return redirect(request.META.get('HTTP_REFERER', ''))
+    if request.session['level'] == 2:
+        CtnrUser.objects.get(ctnr_id=ctnr_pk, user_id=user_pk).delete()
+        return redirect(request.META.get('HTTP_REFERER', ''))
+    else:
+        messages.error(request,
+                      'You do not have permission to perform this action')
+        return redirect(request.META.get('HTTP_REFERER', ''))
 
 
 def update_user(request, ctnr_pk, user_pk, lvl):
-    ctnr_user = CtnrUser.objects.get(ctnr_id=ctnr_pk, user_id=user_pk)
-    if (ctnr_user.level + int(lvl)) not in range(0, 3):
-        return redirect(request.META.get('HTTP_REFERER', ''))
+    if request.session['level'] == 2:
+        ctnr_user = CtnrUser.objects.get(ctnr_id=ctnr_pk, user_id=user_pk)
+        if (ctnr_user.level + int(lvl)) not in range(0, 3):
+            return redirect(request.META.get('HTTP_REFERER', ''))
+        else:
+            ctnr_user.level += int(lvl)
+            ctnr_user.save()
+            return redirect(request.META.get('HTTP_REFERER', ''))
     else:
-        ctnr_user.level += int(lvl)
-        ctnr_user.save()
+        messages.error(request,
+                      'You do not have permission to perform this action')
         return redirect(request.META.get('HTTP_REFERER', ''))
